+++ CHANGELOG.md
@@ -6,4 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Show App version, commit hash and build date in App settings page (#70).
+
 [unreleased]: https://github.com/olivierlacan/keep-a-changelog/